refactor: normalize file naming
