+++ packages/g6/src/stdlib/data/index.ts
@@ -1,4 +1,4 @@
-export * from './mapNodeSize';
-export * from './processParallelEdges';
-export * from './transformV4Data';
-export * from './validateData';
+export * from './map-node-size';
+export * from './process-parallel-edges';
+export * from './transform-v4-data';
+export * from './validate-data';
