+++ packages/g6/src/stdlib/data/map-node-size.ts
