+++ packages/g6/src/stdlib/data/process-parallel-edges.ts
