+++ packages/g6/src/stdlib/data/validate-data.ts
