+++ packages/g6/src/stdlib/item/node/model-rect.ts
