+++ packages/g6/src/stdlib/plugin/edge-bundling/index.ts
