+++ packages/g6/src/stdlib/plugin/index.ts
@@ -1,10 +1,10 @@
-export * from './edgeBundling';
-export * from './edgeFilterLens';
+export * from './edge-bundling';
+export * from './edge-filter-lens';
 export * from './fisheye';
 export * from './grid';
 export * from './history';
 export * from './legend';
-export * from './lodController';
+export * from './lod-controller';
 export * from './menu';
 export * from './minimap';
 export * from './snapline';
